+++ cfgrib/messages.py
@@ -46,7 +46,7 @@ _MARKER = object()
 #
 # No explicit support for MULTI-FIELD at Message level.
 #
-# bindings.codes_grib_multi_support_on()
+bindings.codes_grib_multi_support_on()
 
 
 @attr.attrs()
@@ -62,9 +62,15 @@ class Message(collections.MutableMapping):
     @classmethod
     def from_file(cls, file, offset=None, product_kind=bindings.CODES_PRODUCT_GRIB, **kwargs):
         # type: (T.IO[bytes], int, int, T.Any) -> Message
+        field_in_message = 0
+        if isinstance(offset, tuple):
+            offset, field_in_message = offset
         if offset is not None:
             file.seek(offset)
         codes_id = bindings.codes_handle_new_from_file(file, product_kind)
+        # iterate over multi-fields in the message
+        for _ in range(field_in_message):
+            codes_id = bindings.codes_handle_new_from_file(file, product_kind)
         return cls(codes_id=codes_id, **kwargs)
 
     @classmethod
@@ -190,26 +196,6 @@ class ComputedKeysMessage(Message):
             return super(ComputedKeysMessage, self).__setitem__(item, value)
 
 
-def make_message_schema(message, schema_keys, log=LOG):
-    schema = collections.OrderedDict()
-    for key in schema_keys:
-        bkey = key.encode(message.encoding)
-        try:
-            key_type = bindings.codes_get_native_type(message.codes_id, bkey)
-        except bindings.EcCodesError as ex:
-            if ex.code != bindings.lib.GRIB_NOT_FOUND:  # pragma: no cover
-                log.exception("key %r failed", key)
-            schema[key] = ()
-            continue
-        size = bindings.codes_get_size(message.codes_id, bkey)
-        if key_type == bindings.CODES_TYPE_STRING:
-            length = bindings.codes_get_length(message.codes_id, bkey)
-            schema[key] = (key_type, size, length)
-        else:
-            schema[key] = (key_type, size)
-    return schema
-
-
 @attr.attrs()
 class FileStream(collections.Iterable):
     """Iterator-like access to a filestream of Messages."""
@@ -266,25 +252,19 @@ def compat_create_exclusive(path, *args, **kwargs):
 
 @attr.attrs()
 class FileIndex(collections.Mapping):
+    allowed_protocol_version = '1'
     filestream = attr.attrib(type=FileStream)
     index_keys = attr.attrib(type=T.List[str])
     offsets = attr.attrib(repr=False, type=T.List[T.Tuple[T.Tuple[T.Any, ...], T.List[int]]])
 
     @classmethod
     def from_filestream(cls, filestream, index_keys):
-        # FIXME: using `Message.message_get` with an explicit message schema was a significant
-        #   optimization at some point, due to less calls to the slow CFFI ABI interface.
-        #   This doesn't appear to be reproducible at the moment so the optimisation is
-        #   disabled and we may choose to remove `make_message_schema` altogether.
-        schema = make_message_schema(filestream.first(), index_keys)
         offsets = collections.OrderedDict()
+        count_offsets = {}  # type: T.Dict[int, int]
         for message in filestream:
             header_values = []
-            for key, args in schema.items():
+            for key in index_keys:
                 try:
-                    # if args and not key == 'time':
-                    #     value = message.message_get(key, *args)
-                    # else:
                     value = message[key]
                 except:
                     value = 'undef'
@@ -292,8 +272,17 @@ class FileIndex(collections.Mapping):
                     value = tuple(value)
                 header_values.append(value)
             offset = message.message_get('offset', bindings.CODES_TYPE_LONG)
-            offsets.setdefault(tuple(header_values), []).append(offset)
-        return cls(filestream=filestream, index_keys=index_keys, offsets=list(offsets.items()))
+            if offset in count_offsets:
+                count_offsets[offset] += 1
+                offset_field = (offset, count_offsets[offset])
+            else:
+                count_offsets[offset] = 0
+                offset_field = offset
+            offsets.setdefault(tuple(header_values), []).append(offset_field)
+        self = cls(filestream=filestream, index_keys=index_keys, offsets=list(offsets.items()))
+        # record the index protocol version in the instance so it is dumped with pickle
+        self.index_protocol_version = cls.allowed_protocol_version
+        return self
 
     @classmethod
     def from_indexpath(cls, indexpath):
@@ -327,8 +316,10 @@ class FileIndex(collections.Mapping):
             filestream_mtime = os.path.getmtime(filestream.path)
             if index_mtime >= filestream_mtime:
                 self = cls.from_indexpath(indexpath)
+                allowed_protocol_version = self.allowed_protocol_version
                 if getattr(self, 'index_keys', None) == index_keys and \
-                        getattr(self, 'filestream', None) == filestream:
+                        getattr(self, 'filestream', None) == filestream and \
+                        getattr(self, 'index_protocol_version', None) == allowed_protocol_version:
                     return self
                 else:
                     log.warning("Ignoring index file %r incompatible with GRIB file", indexpath)